Move rename_bika_setup to upgrade step 1.1.9
